Fix excessively high rate of false positives (fixes #15)
The problem was that the locations of the bloom filter bits for a
particular entry were not independent pseudorandom values as they should
be. In particular, if you have two strings that have a FNV collision
(for example, fnv_1a('K3040109') == fnv_1a('K5338215') == 8754595), then
their fnv_1a_b values will also be equal (since fnv_1a_b depends only on
the result of the first hash). Hence, the Bloom filter bit locations for
all of the colliding strings will always be the same:

    var bloom = new BloomFilter(1000, 4)
    bloom.add('K3040109')
    bloom.test('K5338215') // false positive: always returns true

To correctly implement Bloom filter characteristics, the two hash
functions must be independent -- that is, even if two strings have a
collision on the first hash, it is very unlikely that they also have a
collision on the second hash.

I think the simplest way of implementing this with FNV is to use two
different base offsets for the two hash functions. I have done this in
the attached patch. I think this is reasonable because the base offset
is chosen essentially arbitrarily (according to the website
http://www.isthe.com/chongo/tech/comp/fnv/index.html the standard base
offset is a hash of one of the authors' email signature). The seed I set
for the second hash function was generated with
Math.ceil(Math.random() * 0xffffffff) -- in the fine tradition of
https://www.xkcd.com/221/ !

Experimental tests show that this patch brings down the false positive
rate to that which is expected from the theory.