Switch to triple-hashing for improved false positive rate.
This reverts back to my initial use of modified FNV-1a for performance:
instead of re-hashing data for each hash function, we simply perform
another round of FNV-1a to get a second hash function.

We do the same again to get a third, which reduces false positives
substantially compared to double hashing.

One additional fix is that we ensure the second hash is always odd,
which avoids catastrophic situations where it is 0 or divides `m`
exactly, assuming `m` is a power of two.

Fixes #31 and #15.